hitl_langgraph/pyproject.toml
[project]
name = "hitl-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-arcade>=1.3.1",
    "langchain[openai]>=0.3.25",
    "langgraph-swarm>=0.0.11",
    "python-dotenv>=1.1.0",
]
